requirements.txt
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
awscli==1.27.59
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.26.118
botocore==1.29.118
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-default-group-wheel==1.2.2
colorama==0.4.6
comm==0.1.3
command-not-found==0.3
cryptography==40.0.2
dbus-python==1.2.18
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
docutils==0.16
ecdsa==0.18.0
entrypoints==0.4
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.3
ffmpeg==1.4
ffmpeg-python==0.2.0
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.20.2
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==8.10.0
moviepy==1.0.3
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.0
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.19.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytube==12.1.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlite-fts4==1.0.3
sqlite-utils==3.30
stack-data==0.6.2
starlette==0.26.1
streamlit==1.21.0
systemd-python==234
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
youtube-dl==2021.12.17
zipp==3.15.0
